fix: tests + socketaddr dtor + add another masternode list engine initialisation method
