+++ dash-spv-apple-bindings/Cargo.toml
@@ -18,7 +18,7 @@ dash-spv-crypto = { workspace = true, features = ["default"] }
 dash-spv-masternode-processor = { workspace = true, features = ["default"] }
 dash-spv-platform = { workspace = true }
 
-dashcore = { git = "https://github.com/dashpay/rust-dashcore", features = ["std", "secp-recovery", "rand", "signer", "serde", "apple", "message_verification", "quorum_validation"], default-features = false, branch = "feat/ferment" }
+dashcore = { git = "https://github.com/dashpay/rust-dashcore", features = ["std", "secp-recovery", "rand", "signer", "serde", "apple", "message_verification", "quorum_validation", "ed25519-dalek"], default-features = false, branch = "feat/ferment" }
 dash-sdk = { git = "https://github.com/dashpay/platform", branch = "test/ferment", package = "dash-sdk", features = ["apple"] }
 data-contracts = { git = "https://github.com/dashpay/platform", branch = "test/ferment", package = "data-contracts", features = ["apple"] }
 dpp = { git = "https://github.com/dashpay/platform", branch = "test/ferment", package = "dpp", default-features = false, features = ["dash-sdk-features", "apple"] }
