+++ dash-spv-apple-bindings/src/tests/listdiff/core21_testnet.rs
@@ -1,8 +1,7 @@
-use std::ptr::null;
 use std::sync::Arc;
 use dash_spv_crypto::network::ChainType;
 use dash_spv_masternode_processor::block_store::{init_mainnet_store, init_testnet_store};
-use dash_spv_masternode_processor::processing::ProcessingError;
+use dash_spv_masternode_processor::processing::processor::processing_error::ProcessingError;
 use dash_spv_masternode_processor::test_helpers::load_message;
 use dash_spv_masternode_processor::tests::FFIContext;
 use crate::ffi_core_provider::FFICoreProvider;
@@ -123,10 +122,10 @@ pub fn core21_mainnet() {
 pub fn testnet_quorums_70235() {
     let chain = ChainType::TestNet;
     let context = Arc::new(FFIContext::chain_default(chain.clone(), false, init_testnet_store()));
-    let processor = FFICoreProvider::default_processor(Arc::clone(&context), chain.clone());
+    let mut processor = FFICoreProvider::default_processor(Arc::clone(&context), chain.clone());
 
     let message = load_message(chain.identifier(), "MNL_0_530000__70228.dat");
-    let maybe_result = processor.mn_list_diff_result_from_message(&message, true, 70228, false, null());
+    let maybe_result = processor.process_mn_list_diff_result_from_message(&message, None, true);
     match maybe_result {
         Ok(_) |
         Err(ProcessingError::MissingLists(..)) => {},
@@ -135,7 +134,7 @@ pub fn testnet_quorums_70235() {
 
     // let protocol_version = extract_protocol_version_from_filename("QRINFO_1194627_1194629__70235.dat").unwrap_or(70219);
     let message = load_message(chain.identifier(), "QRINFO_1194627_1194629__70235.dat");
-    let maybe_result = processor.qr_info_result_from_message(&message, true, 70235, true, false, null());
+    let maybe_result = processor.process_qr_info_result_from_message(&message, true, true);
     match maybe_result {
         Ok(_) |
         Err(ProcessingError::MissingLists(..)) => {},
