+++ dash-spv-apple-bindings/src/tests/mod.rs
@@ -1,2 +1,2 @@
 pub mod common;
-pub mod listdiff;
+// pub mod listdiff;
