+++ dash-spv-crypto/Cargo.toml
@@ -16,7 +16,7 @@ bip38 = { git = "https://github.com/pankcuf/bip38", rev = "87abd21" }
 blake3 = "1.3.2"
 bls-signatures = { git = "https://github.com/dashpay/bls-signatures", tag = "1.3.5", features = ["legacy", "bip32", "apple", "use_serde" ]  }
 byte = "0.2"
-dashcore = { git = "https://github.com/dashpay/rust-dashcore", features = ["std", "secp-recovery", "rand", "signer", "serde", "apple", "message_verification", "quorum_validation"], default-features = false, branch = "feat/ferment" }
+dashcore = { git = "https://github.com/dashpay/rust-dashcore", features = ["std", "secp-recovery", "rand", "signer", "serde", "apple", "message_verification", "quorum_validation", "ed25519-dalek"], default-features = false, branch = "feat/ferment" }
 dash-spv-keychain = { workspace = true }
 
 ed25519-dalek = "2.1.1"
