+++ dash-spv-masternode-processor/src/tests/hashes/mod.rs
@@ -1,3 +1,2 @@
 pub mod hashes;
 pub mod math;
-pub mod multiple_merkle_hashes;
